added verifications on non existing objects (IDs)

+++ handlers/deletepost.py
@@ -13,10 +13,13 @@ class DeletePost(Handler):
         dbuser = User.get_by_id(int(cookie_uid))
         postid_=self.request.get('postid')
         dbpost = Post.get_by_id(int(postid_))
-        if dbuser.key() == dbpost.user.key():
-            self.render("deletepost.html", post_title=dbpost.subject, postid=postid_)
+        if dbuser and dbpost:
+            if dbuser.key() == dbpost.user.key():
+                self.render("deletepost.html", post_title=dbpost.subject, postid=postid_)
+            else:
+                self.redirect("/message/8")
         else:
-            self.redirect("/message/8")
+            self.redirect("/message/9")
 
     def post(self):
         postid_= self.request.get('postid')
@@ -30,8 +33,11 @@ class DeletePost(Handler):
             dbuser = User.get_by_id(int(cookie_uid))
             if dbuser and (dbuser.HashedPassword==cookie_hash):
                 dbpost = Post.get_by_id(int(postid_))
-                dbpost.delete()
-                self.redirect("/")
+                if dbpost:
+                    dbpost.delete()
+                    self.redirect("/")
+                else:
+                    self.redirect("/message/9")
             # cookie is not valid
             else:
                 self.redirect("/message/4")
